GSIs controller fix
