+++ spec/models/user_spec.rb
@@ -37,7 +37,7 @@ describe User do
     end
   end
 
-  describe 'signed_in_before() method' do
+  describe 'signed_in_before' do
     let(:user) { create(:user) }
     it 'returns true if sign_in_count > 0' do
       user.sign_in_count = 3
@@ -49,6 +49,29 @@ describe User do
       user.sign_in_count = 0
       user.save!
       expect(user.signed_in_before).to be false
+    end
+  end
+
+  describe 'appointments_count' do
+    let (:course1) { create(:course) }
+    let (:course2) { create(:course) }
+    let (:user) { create(:user) }
+
+    it 'returns 0 if the person has no appointments' do
+      expect(user.appointments_count).to eq 0
+    end
+
+    it 'returns 1 if the person has 1 appointment' do
+      user.courses_to_teach << course1
+      user.save!
+      expect(user.appointments_count).to eq 1
+    end
+
+    it 'returns 2 if the person has 2 appointments' do
+      user.courses_to_teach << course1
+      user.courses_to_teach << course2
+      user.save!
+      expect(user.appointments_count).to eq 2
     end
   end
 end
